# Rewrite read_frame(), use AES_CTR for big speed-up