@DLu 2024-01-04
My results: 
 * Old version: 8m29s
 * New version: 0m 9s

This is amazing. Thanks so much.